refactoring classses out
